refactors tide into separate component
