+++ client/src/components/SmallMap.js
@@ -1,0 +1,34 @@
+import { useEffect, useState, useRef } from "react";
+import "@tomtom-international/web-sdk-maps/dist/maps.css";
+import * as tt from "@tomtom-international/web-sdk-maps";
+const ttApiKey = require("../tt-api-key");
+
+export const SmallMap = (props) => {
+  const mapElement = useRef();
+  const [map, setMap] = useState({});
+
+  useEffect(() => {
+    let map = tt.map({
+      key: ttApiKey,
+      container: mapElement.current,
+      center: [props.long, props.lat],
+      zoom: 15,
+    });
+
+    var element = document.createElement("div");
+    element.id = "marker";
+    let marker = new tt.Marker({
+      element: element,
+    })
+      .setLngLat([props.long, props.lat])
+      .addTo(map);
+    setMap(map);
+    return () => map.remove();
+  }, []);
+
+  return (
+    <div>
+      <div ref={mapElement} className="mapDiv" />
+    </div>
+  );
+};
